Add space operations interface, move and rename space operations implementation class
